pyproject.toml
[project]
name = "elastic-qdrant-benchmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "elasticsearch>=9.2.0",
    "embedding-atlas>=0.12.0",
    "polars>=1.35.2",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
    "umap-learn>=0.5.9.post2",
]
